Configure for use with AWS, Remove db role
Add an alternate Vagrantfile that we can use to provision on AWS.

We'll be using RDS for the application database, so there's no need
to provision Postgres servers (AWS handles this for us).

The only caveat is that we now must be sure that a database already
exists for the Django application.