Double VM memory to 1 GiB
In production, we're using a t2.micro (as opposed to a t2.nano). With
Celery, RabbitMQ, and other services running, we were actually able to
exhaust the 512 MiB of memory (specifically, node was failing on `bower
install`).

A local dev environment should mirror production as closely as possible
anyway, so bring the VM's memory size up to 1 GiB.